fix: #70: change double to decimal for tracked hours for more accurate counting tracked hours (#72)
* fix: #70: change double to decimal for tracked hours for more accurate counting tracked hours

* docs(readme): bring test coverage score up to date

* refactor: #70: change cast to decimal to use suffix "m" for decimal

* test: #70: update TotalTrackedTaskHoursCalculatorWithIncompleteHours test

* test: #70: update TotalTrackedTaskHoursCalculatorWithIncompleteHours test

* docs(readme): bring test coverage score up to date

---------

+++ Application/Features/Internal/GetEmployeesTrackedTaskHours/GetEmployeesTrackedTaskHoursResponse.cs
@@ -9,5 +9,5 @@ public class EmployeeTrackedTaskHourDto
 {
     public required long EmployeeId { get; set; }
 
-    public required double TrackedHours { get; set; }
+    public required decimal TrackedHours { get; set; }
 }
